+++ packages/lumx-react/src/components/list/index.ts
@@ -2,4 +2,5 @@ export * from './List';
 export * from './ListItem';
 export * from './ListItemAction';
 export * from './ListDivider';
+export * from './ListSection';
 export * from './ListSubheader';