+++ packages/site-demo/content/product/components/list/react/item-padding.tsx
@@ -1,39 +1,39 @@
 import { classNames } from '@lumx/core/js/utils';
-import { List, ListItem, ListSubheader, Text } from '@lumx/react';
+import { List, ListItem, ListSection, Text } from '@lumx/react';
 
 export default () => (
     <>
         <List>
-            <ListSubheader>Default</ListSubheader>
-
-            <ListItem size="big" className={classNames.background('dark-L6')}>
-                <Text as="p">Two-line item</Text>
-                <Text as="p" color="dark-L2">
-                    Secondary text
-                </Text>
-            </ListItem>
+            <ListSection label="Default">
+                <ListItem size="big" className={classNames.background('dark-L6')}>
+                    <Text as="p">Two-line item</Text>
+                    <Text as="p" color="dark-L2">
+                        Secondary text
+                    </Text>
+                </ListItem>
+            </ListSection>
         </List>
 
         <List itemPadding="big">
-            <ListSubheader>Big padding</ListSubheader>
-
-            <ListItem size="big" className={classNames.background('dark-L6')}>
-                <Text as="p">Two-line item</Text>
-                <Text as="p" color="dark-L2">
-                    Secondary text
-                </Text>
-            </ListItem>
+            <ListSection label="Big padding">
+                <ListItem size="big" className={classNames.background('dark-L6')}>
+                    <Text as="p">Two-line item</Text>
+                    <Text as="p" color="dark-L2">
+                        Secondary text
+                    </Text>
+                </ListItem>
+            </ListSection>
         </List>
 
         <List itemPadding="huge">
-            <ListSubheader>Huge padding</ListSubheader>
-
-            <ListItem size="big" className={classNames.background('dark-L6')}>
-                <Text as="p">Two-line item</Text>
-                <Text as="p" color="dark-L2">
-                    Secondary text
-                </Text>
-            </ListItem>
+            <ListSection label="Huge padding">
+                <ListItem size="big" className={classNames.background('dark-L6')}>
+                    <Text as="p">Two-line item</Text>
+                    <Text as="p" color="dark-L2">
+                        Secondary text
+                    </Text>
+                </ListItem>
+            </ListSection>
         </List>
     </>
 );